[FIX] ShabbatTV
ChatGPT

The provided code seems to be a mix of JavaScript and Liquid syntax for Jekyll.
However, there are syntax errors and inconsistencies. Let's correct it:

```

{% comment %}
This is a Jekyll comment. It won't be rendered in the final HTML output.
{% endcomment %}

<script>
document.addEventListener("click", function(){
    const backscreen = PravdaTV(new THREE.PlaneGeometry({{ page.PravdaSize | default: page.TVSize | default: "5, 2.8" }}), {{ page.TVRotation | default: "0"}});
    backscreen.position.set({{ page.PravdaTV }});

    const screen = ShabbatTV(new THREE.PlaneGeometry({{ page.TVSize | default: "5, 2.8" }}), {{ page.TVRotation | default: "0"}});
    screen.position.set({{ page.ShabbatTV }});
});
</script>

```

> Changes made:
>
> 1. Wrapped the JavaScript code in a `<script>` tag.
> 2. Used Liquid's default filter properly.
> 3. Removed the incorrect usage of `default` in the second instance.
>

> Ensure that `page.PravdaSize`, `page.TVSize`, `page.TVRotation`,
`page.PravdaTV`, and `page.ShabbatTV` are properly defined in your Jekyll front
matter or YAML data files.