Citizens of ISRAEL (اسرائيل),
Shortly before two pm today, the armies of Egypt and Syria opened an offensive against Israel, launching a series of air, armoured and artillery attacks in Sinai and on the Golan Heights. The Israel Defense Forces have entered the fight, and are beating back the assault. The enemy has suffered grave losses.

The rulers of Egypt and Syria have long planned this violation of the cease-fire. Contemptibly, the aggressors are now spreading the falsehood that it was Israel that opened fire. But the responsibility for the renewal of the fighting and for the bloodshed lies with them alone.

Our enemies had hoped to surprise the citizens of Israel on the Day of Atonement, when so many of our people are fasting and worshipping in the synagogues. The aggressors thought that on this day we would not be ready to fight back. We were not caught by surprise.

For several days now our Intelligence Services have been aware that the armies of Egypt and Syria were preparing a joint offensive. Israel Defense Force patrols established that large armed forces were massed in offensive deployment in the vicinity of the Suez Canal and on the Golan Heights. The reports of the patrols confirmed the information already in our hands. Our forces were placed in position to meet the danger.

We are in no doubt that we shall prevail, but we are also convinced that this renewal of Egyptian and Syrian aggression is an act of madness. We did our best to prevent the outbreak. We appealed to quarters with political influence to use it in order to frustrate this infamous move of the Egyptian and Syrian leaders. While there was still time we informed friendly countries of the confirmed information that we had of the plans for an offensive against Israel. We called on them to do the utmost to prevent war, but the Egyptian and Syrian attack has started.

The Israel Defense Forces are ready to repel the enemy's attack. Early this morning a partial call-up of reserves was approved and got under way.

In view of the gravity of the information, I was obliged to call a meeting of the Cabinet on the Day of Atonement. The offensive started while the Cabinet was in session. The Cabinet authorized the IDF to take all action on the battlefront required by the situation to achieve victory. Ministers were authorized to issue the necessary emergency orders.

Citizens of Israel,

This is not the first time that we have been compelled to endure ordeal by battle. I am confident that none among us will fall prey to panic. The mobilization will no doubt cause hardships and interference in the normal course of life and work but we must try to accept these difficulties as we have done in the past, with full understanding. We are called upon to demonstrate responsibility and discipline in our conduct. We must be ready for every burden and sacrifice needed for the defense of our survival, our freedom and our independence. Let us, then, conduct ourselves so as to be worthy of our soldiers of Israel who are valiantly doing their duty in the air, in the armoured forces, on the ships, in the artillery, in all units and services in the outposts, in the villages and settlements, along the lines of fire in all sectors.

We have full confidence in the spirit and the strength of the IDF to overcome the enemy. The victory of the IDF is our certain assurance of life and peace.